ActivityAnalyzer/mask_rcnn/requirements.txt
opencv-python==4.5.1.48
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cu102
torchvision==0.9.1+cu102
torchaudio===0.8.1
matplotlib==3.2.0
natsort==7.1.1
numpy==1.20.2
onnx==1.8.1
onnxruntime==1.7.0
pandas==1.2.4
Pillow==8.2.0
protobuf==3.15.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
typing-extensions==3.7.4.3
cycler==0.10.0
kiwisolver==1.3.1

# NOTE: the pycocotools library is also a requirment, but it requires some of these packages to already be installed.
# Due to the limitations of the requirments file, pycocotools must be installed seperately AFTER executing this file.
# You can install it by running the command "pip install pycocotools-windows"

